JACOBIN-104 Fixed jump logic of IF_ICMPLT instruction

+++ src/exec/run.go
@@ -124,8 +124,8 @@ func runFrame(f *frame) error {
 			val2 := pop(f)
 			val1 := pop(f)
 			if val1 < val2 { // if comp succeeds, next 2 bytes hold instruction index
-				jumpTo := (int(f.meth[f.pc+1]) * 256) + int(f.meth[f.pc+2])
-				f.pc = f.pc + jumpTo - 1 // -1 b/c on the next iteration, pc is bumped by 1
+				jumpTo := (int16(f.meth[f.pc+1]) * 256) + int16(f.meth[f.pc+2])
+				f.pc = f.pc + int(jumpTo) - 1 // -1 b/c on the next iteration, pc is bumped by 1
 			} else {
 				f.pc += 2
 			}
@@ -133,8 +133,8 @@ func runFrame(f *frame) error {
 			val2 := pop(f)
 			val1 := pop(f)
 			if val1 >= val2 { // if comp succeeds, next 2 bytes hold instruction index
-				jumpTo := (int(f.meth[f.pc+1]) * 256) + int(f.meth[f.pc+2])
-				f.pc = f.pc + jumpTo - 1 // -1 b/c on the next iteration, pc is bumped by 1
+				jumpTo := (int16(f.meth[f.pc+1]) * 256) + int16(f.meth[f.pc+2])
+				f.pc = f.pc + int(jumpTo) - 1 // -1 b/c on the next iteration, pc is bumped by 1
 			} else {
 				f.pc += 2
 			}
